move: 이동 enum 코드 파일 분리
Control.java -> Enum.java
Output.java -> import lotto.Control.Reward, lotto.Control.SameNumber 제거
